+++ scaffold/src/components/NoticeIcon/demo/basic.md
@@ -1,0 +1,12 @@
+---
+order: 1
+title: 通知图标
+---
+
+通常用在全局导航上。
+
+````jsx
+import { NoticeIcon } from 'ant-design-pro';
+
+ReactDOM.render(<NoticeIcon count={5} />, mountNode);
+````
