+++ scaffold/src/components/NoticeIcon/index.less
