+++ scaffold/src/layouts/BasicLayout.js
@@ -1,17 +1,88 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Layout, Menu, Icon, Avatar, Dropdown } from 'antd';
+import { Layout, Menu, Icon, Avatar, Dropdown, Tag } from 'antd';
 import DocumentTitle from 'react-document-title';
 import { connect } from 'dva';
 import { Link, routerRedux } from 'dva/router';
+import moment from 'moment';
 import styles from './BasicLayout.less';
 import HeaderSearch from '../components/HeaderSearch';
-import NotificationIcon from '../components/NotificationIcon';
+import NoticeIcon from '../components/NoticeIcon';
 import GlobalFooter from '../components/GlobalFooter';
 import { menus } from '../common/nav';
 
 const { Header, Sider, Content } = Layout;
 const { SubMenu } = Menu;
+
+const data1 = [{
+  key: '1',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/ThXAXghbEsBCCSDihZxY.png',
+  title: '你收到了 14 份新周报',
+  datetime: moment('2017-08-09').fromNow(),
+}, {
+  key: '2',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/OKJXDXrmkNshAMvwtvhu.png',
+  title: '你推荐的 曲妮妮 已通过第三轮面试',
+  datetime: moment('2017-08-08').fromNow(),
+}, {
+  key: '3',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/kISTdvpyTAhtGxpovNWd.png',
+  title: '这种模板可以区分多种通知类型',
+  datetime: moment('2017-08-07').fromNow(),
+  read: true,
+}, {
+  key: '4',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/GvqBnKhFgObvnSGkDsje.png',
+  title: '左侧图标用于区分不同的类型',
+  datetime: moment('2017-08-07').fromNow(),
+}, {
+  key: '5',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/ThXAXghbEsBCCSDihZxY.png',
+  title: '内容不要超过两行字，超出时自动截断',
+  datetime: moment('2017-08-07').fromNow(),
+}];
+
+const data2 = [{
+  key: '1',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/fcHMVNCjPOsbUGdEduuv.jpeg',
+  title: '曲丽丽 评论了你',
+  description: '描述信息描述信息描述信息',
+  datetime: moment('2017-08-07').fromNow(),
+}, {
+  key: '2',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/fcHMVNCjPOsbUGdEduuv.jpeg',
+  title: '朱偏右 回复了你',
+  description: '这种模板用于提醒谁与你发生了互动，左侧放『谁』的头像',
+  datetime: moment('2017-08-07').fromNow(),
+}, {
+  key: '3',
+  avatar: 'https://gw.alipayobjects.com/zos/rmsportal/fcHMVNCjPOsbUGdEduuv.jpeg',
+  title: '标题',
+  description: '这种模板用于提醒谁与你发生了互动，左侧放『谁』的头像',
+  datetime: moment('2017-08-07').fromNow(),
+}];
+
+const data3 = [{
+  key: '1',
+  title: '任务名称',
+  description: '任务需要在 2017-01-12 20:00 前启动',
+  extra: <Tag color="red">马上到期</Tag>,
+}, {
+  key: '2',
+  title: '第三方紧急代码变更',
+  description: '冠霖提交于 2017-01-06，需在 2017-01-07 前完成代码变更任务',
+  extra: <Tag color="red">马上到期</Tag>,
+}, {
+  key: '3',
+  title: '信息安全考试',
+  description: '指派竹尔于 2017-01-09 前完成更新并发布',
+  extra: <Tag color="yellow">已耗时 8 天</Tag>,
+}, {
+  key: '4',
+  title: 'ABCD 版本发布',
+  description: '冠霖提交于 2017-01-06，需在 2017-01-07 前完成代码变更任务',
+  extra: <Tag color="blue">进行中</Tag>,
+}];
 
 class BasicLayout extends React.PureComponent {
   static childContextTypes = {
@@ -150,7 +221,14 @@ class BasicLayout extends React.PureComponent {
               />
               <div className={styles.right}>
                 <HeaderSearch className={`${styles.action} ${styles.search}`} placeholder="站内搜索" />
-                <NotificationIcon className={styles.action} count={currentUser.notifyCount} />
+                <NoticeIcon
+                  className={styles.action}
+                  count={currentUser.notifyCount}
+                >
+                  <NoticeIcon.Tab list={data1}>通知（4）</NoticeIcon.Tab>
+                  <NoticeIcon.Tab list={data2}>消息（4）</NoticeIcon.Tab>
+                  <NoticeIcon.Tab list={data3}>待办（4）</NoticeIcon.Tab>
+                </NoticeIcon>
                 <Dropdown overlay={menu}>
                   <span className={`${styles.action} ${styles.account}`}>
                     <Avatar size="small" className={styles.avatar} src={currentUser.avatar} />
